requirements.txt
certifi==2021.10.8
dash==2.0.0
dash_bootstrap_components==1.0.2
joblib==1.1.0
langdetect==1.0.9
numpy==1.22.1
pandas==1.3.5
plotly==5.5.0
pymongo==4.0.1
python-dotenv==0.19.2
python_dateutil==2.8.2
pytz==2021.3
requests==2.27.1
snscrape==0.4.3.20220107.dev8+g7f88678
yfinance==0.1.69
